Move TruePointFilter class to the Support namespace
