+++ src/Go/Aop/Pointcut/StaticMethodMatcherPointcut.php
@@ -10,7 +10,7 @@ namespace Go\Aop\Pointcut;
 
 use Go\Aop\Pointcut;
 use Go\Aop\PointFilter;
-use Go\Aop\TruePointFilter;
+use Go\Aop\Support\TruePointFilter;
 use Go\Aop\Support\StaticMethodMatcher;
 
 /**
