+++ tests/Go/Aop/Support/TruePointFilterTest.php
@@ -6,7 +6,7 @@
  * @license       http://www.opensource.org/licenses/mit-license.php The MIT License
  */
 
-namespace Go\Aop;
+namespace Go\Aop\Support;
 
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.0 on 2012-11-28 at 15:42:10.
